data-extractor/requirements.txt
streamlit==1.32.0
chromadb==0.4.24
google-generativeai>=0.4.1
langchain==0.1.12
langchain-community==0.0.28
langchain-google-genai==0.0.11
langchain-chroma==0.1.0
langchain-huggingface==0.0.1
sentence-transformers>=2.6.0
watchdog==4.0.0
pandas==2.2.1
openpyxl==3.1.2
pypdf==4.1.0
unstructured==0.12.5
tenacity==8.2.3
python-dotenv==1.0.1
